requirements.txt
langchain-openai
langchain
python-dotenv
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
faiss-cpu
sentence_transformers
pypdf
pymupdf
streamlit
langchain-chroma
langchain_huggingface
arxiv
wikipedia
SQLAlchemy
duckdb
duckduckgo-search
mysql-connector-python
validators==0.28.1
youtube_transcript_api
unstructured
pytube